+++ lib/utils.h
@@ -52,7 +52,7 @@ void shift_w_to_b( int8 *in, uint8 *out, int shift, uint32 size);
 uint8* utils_shift(int16 *img, uint8 *rgb, uint32 w, uint32 h, int sh);
 int16* utils_shift16(int16 *img, int16 *rgb, uint32 w, uint32 h, int sh);
 
-double entropy(int16 *img, uint32 *buf, uint32 w, uint32 h, uint32 bpp);
+double entropy(uint8 *img, uint32 *buf, uint32 w, uint32 h, uint32 bpp);
 
 
 //uint8* 		utils_subband_draw		(Image *img, uint8 *rgb, ColorSpace color, uint32 steps);
